feat: modify target url construction
Rest endpoint may (and is going to) start with the version
(v2.1/servers). Information in the service catalog is most likely also
containing version information (and project_id). In order to construct
the resulting url correctly go through parts of the service catalog
endpoint information and strip it from the rest endpoint url when it
starts with it.