fix: Respect server defaults for pagination
On a server side there might be own preferred defaults set for the page
size which may vary across different services/resources. So far this was
ignored by the sdk and own page size was enforced. Change this to
actually not try to limit the page size unless a dedicated query
parameter is set by the EP itself.