refactor(sdk): Get rid of anyhow in openstack_sdk (#228)
anyhow is great at adding context to the errors, but it is recommended
to use anyhow in client facing tools, while thiserror in libs. Using it
makes it quite ugly to get good errors in the cli. So get rid of it and
instead extend base errors to be self containing all required
information.