feat(sdk): Add api version into RestEndpoint (#241)
* feat(sdk): Add Unsupported version error

* feat(sdk): Add api version into the resource

As a next step after introducing endpoint version handling it is
necessary to add information about the required/minimum version into
eash `RestEndpoint` directly. With this version prefix is being dropped
and we can rely better on the results of version discovery mechanism.

* fix: Stop deduplication in endpoint URL construction

Now that `RestEndpoint` urls are not containing version prefix it is not
necessary to deduplicate URL. It is actually even harmful since BS
endpoint `http://test.com/volumes` and endpoint `volumes/detail` result
in a broken request. Stop doing deduplication what was necessaty before.