fix(pagination): Address new cornercase in pagination (#312)
Access to a new cloud uncovered one interesting case. There is a single
server in the cloud in the error state and Nova returns 404 for access
by name, while it finds it in the list. However, since this is the only
entry in the list pagination tries to get a next page with the marker
and get's back 400 (invalid marker). For the keyset pagination we should
really stop trying fetching next page when no information about the next
page is available. Swift is not using keyset pagination, and so set it
in the `RestEndpoint` properly while implementing reasonable pagination
using w3c `Link` header (thanks to GitLab SDK implementation).