fix(cli): Fix network.port.dns_assignments schema (#329)
dns_assignments in Neutron are not having any schema information. This
leads to ser/deser issues when the fiels is populated (is usually not
the case in devstack, but often with real clouds). Add required
information to the schema so that cli knows it is a complex object
treating it as json value.