feat(cli): Start switch to eyre for error handling (#357)
* feat(cli): Start switch to eyre for error handling

Start switch from anyhow to eyre/color-eyre for error handling in the
cli due to a much more user friendly reporting.

* fix: Support octavia error messages

Octavia errors are having slightly different format. Add extract of
`faultstring` as message.

* feat(cli): Add `report error` integration

color-eyre crate is having a nice feature of automatically suggesting
user to open GitHub issue with the relevant information. Start using it
knowing we would further improve filter for which reports this might be
relevant.