feat: New generated content
Add crate-ci/typos for typo correction

There are lots of places in the code where typos might occur. Those even
squeeze into the openapi and as such into the generated code. On the
Rust side we already use `typos` to check that so we need to also invoke
it when we generate openapi specs themselves.

Change-Id: I9d27391492f9eb10a6b65d4c4589b12db2610813

Changes are triggered by https://review.opendev.org/925448