fix: Change condition in non-terminal data upload
Previously condition for the input data was wrongly evaluating to read
the data from stdin when it is not a terminal or file name was unset. It
makes no sense and stdin should be used when both conditions are true.
'else' branch is looking at the passed file name and still select stdin
when fname is '-'.

Signed-off-by: Artem Goncharov <artem.goncharov@gmail.com>