chore: Update cargo-dist workflow
Regenerate and re-fix workflow for the cargo-dist.  We also want to be
sure binaries are uploaded into the openstack_cli release and not only
in the vX.Y.Z what requires changes in the template again. For now
restore the carg-dist's native workflow and decide later how to ensure
the hardening.