feat(sdk): Add find_by_name method (#505)
It may be desired to find resource by name explicitly rather then guess
whether identifier is id or name in order to save API roundtrip known
not to return data. This is going to be used in the cli to allow
different handling of the path parameters allowing user to explicitly
specify whether the input is name or id.