fix: Auth with project_id and no domain info set
When project_id is set in the auth but no project_domain is present
Keystone refuses to auth since it checks project_name before the
project_id (which is normally sufficient and does not require
domain_infos). When project_id is present in the auth config do not set
project_name.