feat: Address data sensitivity
In the config and in auth call we deal with the sensitive data (password,
token, etc). Ensuring it is not leaked it not so trivial but aparently the
`secrecy` crate is able to serve the needs. Since we need to be able to
serialize password into the Json we can't ensure at the moment that the value
does not appear in logs, but at least it will not when the holding structure
is included in the log.

With this we can remove `(crate)` visibility restriction from config structs.

And while on that also finally implement censoring of the json requests to
prevent sensitive auth data to appear in trace logs