feat(test): Introduce FakeOpenStackClient (#1114)
In order to allow downstream users to test their resources create fake client
that can be used together with mock server. This is a first step. In the next
step all existing tests will be converted to this new fake client. Afterwards
it will be exposed dropping current MockServerClient and MockAsyncServerClient.