chore: Separate structable into separate repository
Creation of openstack_types crate uncovered an issue with the place
where StructTable trait is being defined (orphan rule). In order to
solve this the interface of StructTable should be changed and we also
need to introduce `structable` crate defining this interface. It makes
no sense anymore to host those 2 crates here. So drop them (keeping
history).