chore: Separate structable into separate repository (#1145)
* feat: Introduce dedicated structable crate

It is not possible to implement StructTable trait for Vec since trait
outside of the place where StructTable trait is defined due to the
orphan rules constraint. Reasonable possibility is to rework StructTable
trait to return headers, data row with a function combining those.

* chore: Separate structable into separate repository

Creation of openstack_types crate uncovered an issue with the place
where StructTable trait is being defined (orphan rule). In order to
solve this the interface of StructTable should be changed and we also
need to introduce `structable` crate defining this interface. It makes
no sense anymore to host those 2 crates here. So drop them (keeping
history).