feat: Switch cli to `openstack_types`
Switch CLI to use the response types defined in `openstack_types`. This
is a bigger change that required stepwise changes and manual
regeneration. After this is merged corresponding codegenerator change
would be merged (it is not possible to do it the regular way due to the
interface change in the structable).