feat: Introduce CLI configuration
Prepare for making CLI output configurable using config files (similarly
to the TUI configuration). The config is searched in the
`$XDG_CONFIG_DIR/osc/[config.yaml, views.yaml]`.

This change is only preparing for the next step of passing additional
information from the codegenerator, so it is not feature complete.