fix: Fix role_assignments command
In the cli generation it was assumed that the first element in the
response schema is the object itself. That is used to check whether the
response is list or not. Now the role_assignment schema return `links`
object first breaking that assumption. Extend the check to first look at
the `response_key` that we identified previously.

Change-Id: Ib2cf6e394486ebd0a475ed3d9bba97feefcd5205

Changes are triggered by https://review.opendev.org/950250