chore: Do correct testing (#1247)
It was intended to run tests with the MSRV and not the clippy. Since
clippy format changes between releases it is not correct to check if the
code formatted with newer clippy passed old clippy. It is relevant for
tests though. Sadly we can't even parse the workspace with the 1.76 as
declared supported by openstack_sdk.