feat: Implement json pointer extractor for the column (#1265)
Add possibility to specify `json_path` as the column configuration in
the configuration file to output only the data under the path. This is
helpful i.e. when listing servers not to show the whole flavor
information, but only show the flavor name using the `/original_name`
json path expression. It is inly applied in the listing operations and
no `-o wide` specified.