feat: Improve service discovery (hopefully) (#1274)
There are lot of possible service and service catalog misconfiguration.
Introduce a dedicated document describing possible problems, add
references in the warnings or error messages and add few more checks.

Since it is not possible to verify that the service url from discovery
url is working (thanks Glance) it is only possible to do some more
"it does look suspicious" checks with warnings and suggestions how to
deal with that.

Current keystoneauth/openstacksdk are not strictly following the
official version discovery process and may actually exit discovery
earlier than this project does exposing difference in behavior.