chore: Add dedicated installer script
We need to get rid of cargo-dist asap since even windows build jobs are
starting to fail (due to GH dropping certain runners). 'cargo-dist' is a
great tool but is abandoned due to the bankruptcy. In the last couple of
month we started getting issues related to lack of updates there, so
prepare the drop.