feat: Make image.locations.metadata not required (#1318)
Glance defines metadata property of the image location as mandatory
while in devstack it is not present. Hardcode a dirty fix for the
schema.

In addition to that in the meanwhile another schema in nova has been
added, so are the changes.

Change-Id: I15f2f0a543174eccbeb5c0d4da9e5222ce1e5d7a
Signed-off-by: Artem Goncharov <artem.goncharov@gmail.com>

Changes are triggered by https://review.opendev.org/953698