feat: Introduce authentication helper
Not always all authentication data can be provided in the config (totp
token, jwt, ...). Till now a dialoger crate was used to interactively
ask for the missing data. This works well in cli, but is absolutely
unsiutable in TUI. Intriduce an AuthHelper trait that can be used as a
sort of callback functionality to allow client to provide necessary
functionality.