feat: Allow cli cloud config from env vars
Apparently there is a real demand for connection configuration using
environment variables. Most likely this is never going to die against
our expectations and hopes. So add support for that, but in difference
to the python world deliberately prevent merging of clouds.yaml data
with the environment variables to prevent unexpected things.