feat: OpenAPI: Fix arrays without `items` (#1348)
As explained in the documentation `arrays` have an `items` section, and
arrays that can have arbitrary values are defined like this [1]:

```
type: array
items: {}
```

In the cinder schemas we have a couple of cases where we are defining an
`array` without `items`, all other arrays in the openapi codebase seem
to be right.

In this patch we fix these issues and introduce code in the
`_sanitize_schema` to detect future cases of this issue and fix the
schema automatically for the output.

[1]: https://swagger.io/docs/specification/v3_0/data-models/data-types/#mixed-type-arrays

Change-Id: I49cadcccf64ed9699a4b41674109b4afbfa1e106

Changes are triggered by https://review.opendev.org/951886