feat: Enable `OS_CLIENT_CONFIG_PATH` environment variable (#1419)
In a way similar to `KUBECONFIG` environment variable add possibility to
predefine locations at which configurations are searched and all merged
into the single "virtual" config.

```
export
OS_CLIENT_CONFIG_PATH=c1.yaml:~/.config/openstack/c3.yaml:/etc/openstack/secure.yaml"
```

When entry point to the file it is tried to be parsed as is. When an
entry is a directory a regular clouds.yaml/secure.yaml files are being
searched in it.